Chapter One, Question Four *Fix

+++ src/chapter_one/Question_Four.java
@@ -6,7 +6,7 @@ public class Question_Four {
         num1 = 75.35;
         num2 = -35.56;
         num3 = 15.76;
-        double average =  (double) (num1 + num2 + num3) /3;
+        double average =  (num1 + num2 + num3) /3;
         System.out.println("Num1 is " + num1 + ", num2 is " + num2 + ", num3 is " + num3 + " and average is " + average);
     }
 }
